+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on Keep a Changelog (https://keepachangelog.com), and this p
 ## [Unreleased] - 2026-02-02
 ### Added
 - Add `DIM` support in Apple II Basic. Add relevant unit tests.
+- `CLEAR` support for arrays. Add relevant unit tests.
 
 ## [Unreleased] - 2026-01-30
 ### Added